maybe runnable added noti for self materialization service
