+++ app/views/roster_notification_mailer/added_to_group_email.html.erb
@@ -1,4 +1,5 @@
 <%= t('mailer.roster_added_to_group_email_body',
+      name: @username,
       rosterable_type: @rosterable.class.name,
       rosterable_title: @rosterable.title,
       rosterable_link: link_to(t('basics.link'),
